refactor: table 패키지 정리
