+++ src/main/java/kitchenpos/table/dao/JdbcTemplateOrderTableDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.table.dao;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
